Update Rust detection and binding generation
This commit updates the build system to automatically detect cargo and
enable/disable _base64 without needing to pass a flag.

It also updates cpython-sys to use a hand written header (which is what
Linux seems to do) and splits off the parser bindings to be handled in
the future (since the files are included differently).